Improved ancestor display in debugger - added invocation ID.
Fixed prefix operator \+ definition to be "\\+" instead of "\+".
Moved p, q, and r definitions to simple_test.pl.
Fixed wam_setup_trace_call to handle target_ftor_ofst param properly, adding TAG_ATM tag to it to create target_ftor value for use in allocating a structure.
Fixed wam_complete_call_or_execute to check that the input predicate is properly defined (e.g. not an empty dynamic predicate).
Wrapped wam() in a try/catch to handle errors with wamExit to ensure that wam timings are halted when an exception is thrown.