Added reducer.pl, sendmore.pl, and zebra.pl bench tests. Added wam tracing.
The zebra.pl test fails due to a problem in the handling of unification.
Extended wam.js and console_test.js to support tracing/debugging WAM instruction execution. This is invoked from the prolog debugger at a CALL port interaction using the new 'm', 'x', 'y', or 'z' commands.
Added decode_instruction function to decode (create a string representation of) an instruction.
Changed predicate_functor to use '.' as the functor for a list.
Fixed predicate_arg to accept a list as the first argument.