Fixed bug in unification. Fixed debugger ancestors to 'freeze'.
Fixed unify() in wam.js to clear the PDL before returning - leaving unprocessed
items on the PDL was causing the zebra test to fail when a subsequen call to unify
would run the remnant unification (now irrelevant) which would cause a variable
to be bound that should not have been bound.
Changed '$trace_push_info' to create string representation of new Goal so that
any subsequent bindings of variables will not affect the display of that Goal
in Ancestor displays.