Included source for examples in documentation HTML.
Extended template.pl to optionally escape '<' and to do simple syntax
coloring for prolog and javascript.
Fixed but in framework.index.template invocation of 'Content'.
Extended index.pl to use new activate_parents/0 predicate to handle case
where the foo.html file being displayed is for an item in the sidenav menu
that is not at the top level of the menu.