Fixed choicepoint and state.trace_call setting.
Fixed create_choicepoint to include TC and TI slots - it was creating a choicepoint that was smaller than the choicepoint created in wam1.
Fixed setting of state.trace_call to always be a Javascript string - it was being set to the Prolog tagged ID of a string in several cases. Also fixed setting of choicepoint TC slot to be the Prolog tagged ID of a string in try_me_else case.
Extended wam_trace_call_or_execute to check that 'functor' does not start with 'debugger:' nor 'system:$trace'.