Fix to compile explicit ':'. Added data_predicates:labelled_values/3.
The ':' fix was to the clause of wam_compiler:compile_body_arg/11 that handles the general arg-as-structure case with MetaArgType of 0..9 or '^'. This fix supports 'call(Module : Goal)'.
The labelled_values/3 predicate returns a key/value list of values defined for the specified data prefix and ID: labelled_values(Module:Prefix, ID, Values).