Added hash and binary search tables for switch instructions.
Fixed FAIL_ADDRESS to be 1000000 (was mistyped as 100000 in someplaces).
Extended wam_index_clauses1/2 to skip 'degenerate' sequences (a sequence of sequences is degenerate if all of the sequences are singletons).
Changed test/3 to be dynamic in web_test_utility.pl to allow for the use of clause/2 on it (since dynamic and public are currently the same flag in ProscriptLS).